# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Smart Farming API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CropScan {
  id: number;
  /** @nullable */
  cropType?: string | null;
  /** @nullable */
  growthStage?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  healthStatus?: string | null;
  /** @nullable */
  diseaseDetected?: string | null;
  /** @nullable */
  nutrientDeficiency?: string | null;
  /** @nullable */
  yieldPredictionKg?: number | null;
  /** @nullable */
  harvestDaysRemaining?: number | null;
  /** @nullable */
  harvestWindow?: string | null;
  /** @nullable */
  confidence?: number | null;
  /** @nullable */
  analysisNotes?: string | null;
  analyzed: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface CreateCropScanBody {
  /** @nullable */
  cropType?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}

export interface CropAnalysis {
  cropType: string;
  growthStage: string;
  healthStatus: string;
  /** @nullable */
  diseaseDetected?: string | null;
  /** @nullable */
  nutrientDeficiency?: string | null;
  yieldPredictionKg: number;
  harvestDaysRemaining: number;
  harvestWindow: string;
  confidence: number;
  analysisNotes: string;
  recommendations: string[];
}

export interface SoilData {
  id: number;
  cropScanId: number;
  /** @nullable */
  phLevel?: number | null;
  /** @nullable */
  moisturePercent?: number | null;
  /** @nullable */
  nitrogenPpm?: number | null;
  /** @nullable */
  phosphorusPpm?: number | null;
  /** @nullable */
  potassiumPpm?: number | null;
  /** @nullable */
  organicMatterPercent?: number | null;
  createdAt: string;
}

export interface CreateSoilDataBody {
  cropScanId: number;
  /** @nullable */
  phLevel?: number | null;
  /** @nullable */
  moisturePercent?: number | null;
  /** @nullable */
  nitrogenPpm?: number | null;
  /** @nullable */
  phosphorusPpm?: number | null;
  /** @nullable */
  potassiumPpm?: number | null;
  /** @nullable */
  organicMatterPercent?: number | null;
}

export interface ClimateData {
  id: number;
  cropScanId: number;
  /** @nullable */
  temperatureCelsius?: number | null;
  /** @nullable */
  humidityPercent?: number | null;
  /** @nullable */
  rainfallMm?: number | null;
  /** @nullable */
  windSpeedKmh?: number | null;
  /** @nullable */
  sunlightHours?: number | null;
  createdAt: string;
}

export interface CreateClimateDataBody {
  cropScanId: number;
  /** @nullable */
  temperatureCelsius?: number | null;
  /** @nullable */
  humidityPercent?: number | null;
  /** @nullable */
  rainfallMm?: number | null;
  /** @nullable */
  windSpeedKmh?: number | null;
  /** @nullable */
  sunlightHours?: number | null;
}

export interface Recommendation {
  id: number;
  cropScanId: number;
  type: string;
  message: string;
  priority: string;
  createdAt: string;
}

export interface ChatMessageBody {
  message: string;
  /** @nullable */
  cropScanId?: number | null;
  /** @nullable */
  cropType?: string | null;
}

export interface ChatResponse {
  reply: string;
  suggestions: string[];
}

export interface CropStatItem {
  cropType: string;
  count: number;
  avgYieldKg: number;
  avgHealthScore: number;
}

export interface DashboardSummary {
  totalScans: number;
  analyzedScans: number;
  healthyCount: number;
  diseasedCount: number;
  avgYieldKg: number;
  avgHarvestDays: number;
  cropBreakdown: CropStatItem[];
}

export type ListCropScansParams = {
  cropType?: string;
  limit?: number;
  offset?: number;
};

export type ListSoilRecordsParams = {
  cropScanId?: number;
};

export type ListClimateRecordsParams = {
  cropScanId?: number;
};

export type ListRecommendationsParams = {
  cropScanId?: number;
};