# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Smart Farming API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all crop scans for the current user
 */
export const ListCropScansQueryParams = zod.object({
  cropType: zod.coerce.string().optional(),
  limit: zod.coerce.number().optional(),
  offset: zod.coerce.number().optional(),
});

export const ListCropScansResponseItem = zod.object({
  id: zod.number(),
  cropType: zod.string().nullish(),
  growthStage: zod.string().nullish(),
  imageUrl: zod.string().nullish(),
  healthStatus: zod.string().nullish(),
  diseaseDetected: zod.string().nullish(),
  nutrientDeficiency: zod.string().nullish(),
  yieldPredictionKg: zod.number().nullish(),
  harvestDaysRemaining: zod.number().nullish(),
  harvestWindow: zod.string().nullish(),
  confidence: zod.number().nullish(),
  analysisNotes: zod.string().nullish(),
  analyzed: zod.boolean(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const ListCropScansResponse = zod.array(ListCropScansResponseItem);

/**
 * @summary Create a new crop scan
 */
export const CreateCropScanBody = zod.object({
  cropType: zod.string().nullish(),
  imageUrl: zod.string().nullish(),
});

/**
 * @summary Get a specific crop scan
 */
export const GetCropScanParams = zod.object({
  id: zod.coerce.number(),
});

export const GetCropScanResponse = zod.object({
  id: zod.number(),
  cropType: zod.string().nullish(),
  growthStage: zod.string().nullish(),
  imageUrl: zod.string().nullish(),
  healthStatus: zod.string().nullish(),
  diseaseDetected: zod.string().nullish(),
  nutrientDeficiency: zod.string().nullish(),
  yieldPredictionKg: zod.number().nullish(),
  harvestDaysRemaining: zod.number().nullish(),
  harvestWindow: zod.string().nullish(),
  confidence: zod.number().nullish(),
  analysisNotes: zod.string().nullish(),
  analyzed: zod.boolean(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Delete a crop scan
 */
export const DeleteCropScanParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Run AI analysis on crop scan (identification, health, yield, harvest prediction)
 */
export const AnalyzeCropScanParams = zod.object({
  id: zod.coerce.number(),
});

export const AnalyzeCropScanResponse = zod.object({
  cropType: zod.string(),
  growthStage: zod.string(),
  healthStatus: zod.string(),
  diseaseDetected: zod.string().nullish(),
  nutrientDeficiency: zod.string().nullish(),
  yieldPredictionKg: zod.number(),
  harvestDaysRemaining: zod.number(),
  harvestWindow: zod.string(),
  confidence: zod.number(),
  analysisNotes: zod.string(),
  recommendations: zod.array(zod.string()),
});

/**
 * @summary List soil data records
 */
export const ListSoilRecordsQueryParams = zod.object({
  cropScanId: zod.coerce.number().optional(),
});

export const ListSoilRecordsResponseItem = zod.object({
  id: zod.number(),
  cropScanId: zod.number(),
  phLevel: zod.number().nullish(),
  moisturePercent: zod.number().nullish(),
  nitrogenPpm: zod.number().nullish(),
  phosphorusPpm: zod.number().nullish(),
  potassiumPpm: zod.number().nullish(),
  organicMatterPercent: zod.number().nullish(),
  createdAt: zod.string(),
});
export const ListSoilRecordsResponse = zod.array(ListSoilRecordsResponseItem);

/**
 * @summary Add soil data for a crop scan
 */
export const CreateSoilDataBody = zod.object({
  cropScanId: zod.number(),
  phLevel: zod.number().nullish(),
  moisturePercent: zod.number().nullish(),
  nitrogenPpm: zod.number().nullish(),
  phosphorusPpm: zod.number().nullish(),
  potassiumPpm: zod.number().nullish(),
  organicMatterPercent: zod.number().nullish(),
});

/**
 * @summary List climate data records
 */
export const ListClimateRecordsQueryParams = zod.object({
  cropScanId: zod.coerce.number().optional(),
});

export const ListClimateRecordsResponseItem = zod.object({
  id: zod.number(),
  cropScanId: zod.number(),
  temperatureCelsius: zod.number().nullish(),
  humidityPercent: zod.number().nullish(),
  rainfallMm: zod.number().nullish(),
  windSpeedKmh: zod.number().nullish(),
  sunlightHours: zod.number().nullish(),
  createdAt: zod.string(),
});
export const ListClimateRecordsResponse = zod.array(
  ListClimateRecordsResponseItem,
);

/**
 * @summary Add climate data for a crop scan
 */
export const CreateClimateDataBody = zod.object({
  cropScanId: zod.number(),
  temperatureCelsius: zod.number().nullish(),
  humidityPercent: zod.number().nullish(),
  rainfallMm: zod.number().nullish(),
  windSpeedKmh: zod.number().nullish(),
  sunlightHours: zod.number().nullish(),
});

/**
 * @summary List recommendations
 */
export const ListRecommendationsQueryParams = zod.object({
  cropScanId: zod.coerce.number().optional(),
});

export const ListRecommendationsResponseItem = zod.object({
  id: zod.number(),
  cropScanId: zod.number(),
  type: zod.string(),
  message: zod.string(),
  priority: zod.string(),
  createdAt: zod.string(),
});
export const ListRecommendationsResponse = zod.array(
  ListRecommendationsResponseItem,
);

/**
 * @summary Send a message to the farming AI chatbot
 */
export const SendChatMessageBody = zod.object({
  message: zod.string(),
  cropScanId: zod.number().nullish(),
  cropType: zod.string().nullish(),
});

export const SendChatMessageResponse = zod.object({
  reply: zod.string(),
  suggestions: zod.array(zod.string()),
});

/**
 * @summary Get overall dashboard stats and summaries
 */
export const GetDashboardSummaryResponse = zod.object({
  totalScans: zod.number(),
  analyzedScans: zod.number(),
  healthyCount: zod.number(),
  diseasedCount: zod.number(),
  avgYieldKg: zod.number(),
  avgHarvestDays: zod.number(),
  cropBreakdown: zod.array(
    zod.object({
      cropType: zod.string(),
      count: zod.number(),
      avgYieldKg: zod.number(),
      avgHealthScore: zod.number(),
    }),
  ),
});

/**
 * @summary Get the most recent crop scans with analysis
 */
export const GetRecentScansResponseItem = zod.object({
  id: zod.number(),
  cropType: zod.string().nullish(),
  growthStage: zod.string().nullish(),
  imageUrl: zod.string().nullish(),
  healthStatus: zod.string().nullish(),
  diseaseDetected: zod.string().nullish(),
  nutrientDeficiency: zod.string().nullish(),
  yieldPredictionKg: zod.number().nullish(),
  harvestDaysRemaining: zod.number().nullish(),
  harvestWindow: zod.string().nullish(),
  confidence: zod.number().nullish(),
  analysisNotes: zod.string().nullish(),
  analyzed: zod.boolean(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const GetRecentScansResponse = zod.array(GetRecentScansResponseItem);

/**
 * @summary Get stats broken down by crop type
 */
export const GetCropStatsResponseItem = zod.object({
  cropType: zod.string(),
  count: zod.number(),
  avgYieldKg: zod.number(),
  avgHealthScore: zod.number(),
});
export const GetCropStatsResponse = zod.array(GetCropStatsResponseItem);