# lib/api-zod/src/generated/types/chatMessageBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Smart Farming API
 * OpenAPI spec version: 0.1.0
 */

export interface ChatMessageBody {
  message: string;
  /** @nullable */
  cropScanId?: number | null;
  /** @nullable */
  cropType?: string | null;
}