# lib/api-zod/src/generated/types/climateData.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Smart Farming API
 * OpenAPI spec version: 0.1.0
 */

export interface ClimateData {
  id: number;
  cropScanId: number;
  /** @nullable */
  temperatureCelsius?: number | null;
  /** @nullable */
  humidityPercent?: number | null;
  /** @nullable */
  rainfallMm?: number | null;
  /** @nullable */
  windSpeedKmh?: number | null;
  /** @nullable */
  sunlightHours?: number | null;
  createdAt: string;
}