# lib/api-zod/src/generated/types/createSoilDataBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Smart Farming API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateSoilDataBody {
  cropScanId: number;
  /** @nullable */
  phLevel?: number | null;
  /** @nullable */
  moisturePercent?: number | null;
  /** @nullable */
  nitrogenPpm?: number | null;
  /** @nullable */
  phosphorusPpm?: number | null;
  /** @nullable */
  potassiumPpm?: number | null;
  /** @nullable */
  organicMatterPercent?: number | null;
}