# lib/api-zod/src/generated/types/cropScan.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Smart Farming API
 * OpenAPI spec version: 0.1.0
 */

export interface CropScan {
  id: number;
  /** @nullable */
  cropType?: string | null;
  /** @nullable */
  growthStage?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  healthStatus?: string | null;
  /** @nullable */
  diseaseDetected?: string | null;
  /** @nullable */
  nutrientDeficiency?: string | null;
  /** @nullable */
  yieldPredictionKg?: number | null;
  /** @nullable */
  harvestDaysRemaining?: number | null;
  /** @nullable */
  harvestWindow?: string | null;
  /** @nullable */
  confidence?: number | null;
  /** @nullable */
  analysisNotes?: string | null;
  analyzed: boolean;
  createdAt: string;
  updatedAt: string;
}