# lib/api-zod/src/generated/types/cropStatItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Smart Farming API
 * OpenAPI spec version: 0.1.0
 */

export interface CropStatItem {
  cropType: string;
  count: number;
  avgYieldKg: number;
  avgHealthScore: number;
}