# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Smart Farming API
 * OpenAPI spec version: 0.1.0
 */
import type { CropStatItem } from "./cropStatItem";

export interface DashboardSummary {
  totalScans: number;
  analyzedScans: number;
  healthyCount: number;
  diseasedCount: number;
  avgYieldKg: number;
  avgHarvestDays: number;
  cropBreakdown: CropStatItem[];
}