# lib/api-zod/src/generated/types/recommendation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Smart Farming API
 * OpenAPI spec version: 0.1.0
 */

export interface Recommendation {
  id: number;
  cropScanId: number;
  type: string;
  message: string;
  priority: string;
  createdAt: string;
}